Add password reset functionality and update app deployment configuration
Introduce password reset feature with email token verification and backend changes for production deployment using Gunicorn, alongside frontend updates to serve static files and display a landing page.

+++ models.py
@@ -83,7 +83,7 @@ class Pick(db.Model):
     home_team = db.Column(db.String, nullable=False)
     game_date = db.Column(db.String, nullable=False)
     side = db.Column(db.String, nullable=False)
-    line = db.Column(db.Integer, default=-110)
+    line = db.Column(db.Float, default=-110)
     edge_pct = db.Column(db.Float, nullable=False)
     model_confidence = db.Column(db.Float, nullable=False)
     result = db.Column(db.String, default='pending')
